feat(fallthough/ptr): support PTR and fallthrough
The plugin already supported fallthrough, but it was done all the time,
and there was no configuration option. This brings in a configuration
option that is done the same way as other CoreDNS plugins that allow for
fallthrough. Thus, if you were relying on fallthrough, you'll need to
manually add a `fallthrough` directive to your Corefile.

It also brings support for reverse DNS without having to manually insert
`PTR` records into the database. In this situation when a `PTR` request
is received, and the `reverse` directive is in the configuration, it
will search for existing `A` records and report back the hostnames
appropriately.

Doing that right required doing some other weird things like properly
reporting SOA when records aren't found, so that's in there, I think. To
make this work, after you've done `pdsql [driver] [dsn]` in your
Corefile, add the list of domains that you're supposed to be an
authority for. It defaults to `.`, so don't worry if you don't have this
set yet.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>