chore(gorm): modernize gorm
This was still using a very old version of gorm, which meant that for
some reason it would always try to redo the schema when it encounted a
database created by a newer version of gorm. This updates the dependency
to use `gorm.io/gorm`, brings in the drivers for the three most common
databases, and updates the schema slightly to ensure that it won't try
to migrate the existing schemas.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>