fix(case-sensitivity): DNS should be case insensitive
This makes it so queries are case insensitive by always lowercasing
requests, as is the general convention. However, because it's not also
calling `lower(records.name)` on the database, you need to make sure
that your database also has the records in lowercase.

The reason for not calling `lower(records.name)` on the database too is
because it will cause issues for indexing on many databases (i.e. it
will cause the database to not use the index if you have the
`records.name` column indexed).

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>